some changes on Grade related classes
