--- eSims/eSims/appsettings.json
@@ -1,20 +1,0 @@
-{
-  
-    "eSimsDatabaseSettings": {
-			"ProfessorsCollectionName": "Professor",
-			"GradesCollectionName": "Grades",
-			"AttendanceCollectionName": "Attendance",
-			"SubjectsCollectionName": "Subjects",
-			"StudentsCollectionName": "Students",
-      "ConnectionString": "mongodb://localhost:27017",
-      "DatabaseName": "eSims"
-    },
-    "Logging": {
-      "LogLevel": {
-        "Default": "Information",
-        "Microsoft": "Warning",
-        "Microsoft.Hosting.Lifetime": "Information"
-      }
-    },
-    "AllowedHosts": "*"
-  }
